created seperate code files for the checkout summary section and payment section of the checkout page. Moved all previous code for the cart summary section to the ordersummary.js file. imported the file to the checkout.js page and ran the renderOrder function to generate the cart summary.
